优化依赖注入模块的结构
