+++ src/plugins/check_in/plugins/check_in_fitness.py
@@ -2,8 +2,8 @@ from nonebot.params import Arg, Depends
 from nonebot.plugin import PluginMetadata
 from nonebot.typing import T_State
 
-from src.utils.annotated import AsyncSession, UserInfo
-from src.utils.helpers import get_plaintext_content, parse_str
+from src.utils.annotated import AsyncSession, PlainTextArgs, UserInfo
+from src.utils.helpers import parse_str
 
 from .. import check_in
 from ..helpers import ensure_user
@@ -22,9 +22,7 @@ fitness_cmd = check_in.command("fitness", aliases={"健身打卡"})
 
 
 @fitness_cmd.handle()
-async def handle_first_message(
-    state: T_State, content: str = Depends(get_plaintext_content)
-):
+async def handle_first_message(state: T_State, content: PlainTextArgs):
     state["content"] = content
