+++ src/plugins/check_in/plugins/check_in_history.py
@@ -12,8 +12,8 @@ from nonebot.plugin import PluginMetadata
 from nonebot.typing import T_State
 from sqlalchemy import func, select
 
-from src.utils.annotated import AsyncSession, UserInfo
-from src.utils.helpers import get_plaintext_content, parse_str
+from src.utils.annotated import AsyncSession, PlainTextArgs, UserInfo
+from src.utils.helpers import parse_str
 
 from .. import check_in
 from ..helpers import ensure_user
@@ -38,9 +38,7 @@ history_cmd = check_in.command("history", aliases={"打卡历史"})
 
 
 @history_cmd.handle()
-async def handle_first_message(
-    state: T_State, content: str = Depends(get_plaintext_content)
-):
+async def handle_first_message(state: T_State, content: PlainTextArgs):
     state["content"] = content
