+++ src/utils/helpers.py
@@ -1,23 +1,14 @@
 import random
 from collections.abc import Sequence
 from datetime import timedelta
-from typing import cast
 
-from nonebot.adapters import Bot, Event, Message, MessageSegment
+from nonebot.adapters import Message
 from nonebot.adapters.onebot.v11 import Bot as OneBotV11Bot
-from nonebot.adapters.onebot.v11 import GroupMessageEvent as OneBotV11GroupMessageEvent
-from nonebot.adapters.onebot.v11 import Message as OneBotV11Message
 from nonebot.adapters.onebot.v12 import Bot as OneBotV12Bot
-from nonebot.adapters.onebot.v12 import (
-    ChannelMessageEvent as OneBotV12ChannelMessageEvent,
-)
-from nonebot.adapters.onebot.v12 import GroupMessageEvent as OneBotV12GroupMessageEvent
-from nonebot.adapters.onebot.v12 import Message as OneBotV12Message
 from nonebot.exception import ActionFailed
 from nonebot.matcher import Matcher
-from nonebot.params import Arg, CommandArg
+from nonebot.params import Arg
 from nonebot.typing import T_State
-from pydantic import BaseModel
 
 from .typing import Expression_T
 
@@ -120,7 +111,7 @@ def timedelta_to_chinese(timedelta: timedelta) -> str:
 
 
 async def get_nickname(
-    bot: Bot,
+    bot: OneBotV11Bot | OneBotV12Bot,
     user_id: str,
     group_id: str | None = None,
     guild_id: str | None = None,
@@ -145,7 +136,7 @@ async def get_nickname(
         # 如果不在群里的话(因为有可能会退群)
         msg = await bot.get_stranger_info(user_id=int(user_id))
         return msg["nickname"]
-    elif isinstance(bot, OneBotV12Bot):
+    else:
         if group_id:
             try:
                 msg = await bot.get_group_member_info(
@@ -184,114 +175,3 @@ async def get_nickname(
             return user["user_name"]
         except ActionFailed:
             pass
-
-
-class MentionedUser(BaseModel):
-    """提及的用户信息"""
-
-    id: str
-    segment: MessageSegment
-
-
-async def get_mentioned_user(args: Message = CommandArg()) -> MentionedUser | None:
-    """获取提到的用户信息"""
-    if isinstance(args, OneBotV11Message) and (at := args["at"]):
-        at = at[0]
-        at = cast(MessageSegment, at)
-        return MentionedUser(id=at.data["qq"], segment=at)
-    if isinstance(args, OneBotV12Message) and (mention := args["mention"]):
-        mention = mention[0]
-        mention = cast(MessageSegment, mention)
-        return MentionedUser(id=mention.data["user_id"], segment=mention)
-
-
-async def get_platform(bot: Bot) -> str | None:
-    """获取平台"""
-    if isinstance(bot, OneBotV11Bot):
-        return "qq"
-    elif isinstance(bot, OneBotV12Bot):
-        return bot.platform
-
-
-class UserInfo(BaseModel, frozen=True):
-    """确定一个用户所需的信息"""
-
-    platform: str
-    user_id: str
-
-
-async def get_user_info(bot: OneBotV11Bot | OneBotV12Bot, event: Event) -> UserInfo:
-    """获取用户信息"""
-    if isinstance(bot, OneBotV11Bot):
-        platform = "qq"
-    else:
-        platform = bot.platform
-
-    user_id = str(event.get_user_id())
-
-    return UserInfo(platform=platform, user_id=user_id)
-
-
-class GroupInfo(BaseModel, frozen=True):
-    """确定一个群或频道所需信息"""
-
-    platform: str
-    group_id: str
-    guild_id: str
-    channel_id: str
-
-    @property
-    def detail_type(self) -> str:
-        """根据是否有 group_id 判断是群还是频道"""
-        if self.group_id:
-            return "group"
-        return "channel"
-
-    @property
-    def send_message_args(self) -> dict[str, str]:
-        """发送消息所需数据"""
-        return {
-            "group_id": self.group_id,
-            "channel_id": self.channel_id,
-            "guild_id": self.guild_id,
-        }
-
-
-async def get_group_info(
-    bot: OneBotV11Bot | OneBotV12Bot,
-    event: OneBotV11GroupMessageEvent
-    | OneBotV12GroupMessageEvent
-    | OneBotV12ChannelMessageEvent,
-) -> GroupInfo:
-    """获取群号或频道号信息"""
-    if isinstance(bot, OneBotV11Bot):
-        platform = "qq"
-    else:
-        platform = bot.platform
-
-    if isinstance(event, OneBotV11GroupMessageEvent):
-        group_id = str(event.group_id)
-        guild_id = ""
-        channel_id = ""
-    elif isinstance(event, OneBotV12GroupMessageEvent):
-        group_id = event.group_id
-        guild_id = ""
-        channel_id = ""
-    else:
-        group_id = ""
-        guild_id = event.guild_id
-        channel_id = event.channel_id
-
-    return GroupInfo(
-        platform=platform,
-        group_id=group_id,
-        channel_id=channel_id,
-        guild_id=guild_id,
-    )
-
-
-async def get_plaintext_content(args: Message = CommandArg()) -> str | None:
-    """获取纯文本命令参数"""
-    if content := args["text"]:
-        if message_str := content.extract_plain_text().strip():
-            return message_str
